chore: merge optional and required parameters
the parameters can be thought of as parameters to the helmfile. We set them in
the environment of the command running helmfile. They are flat in the end so
there should only be one value for each parameter. This is easy to enforce
if they are stored in a map with the key being the name