chore: add parameter validation mechanism (#446)
Validator is a function returning an error if given value is invalid. Using higher-order
functions we can create reusable validators like OneOf. If we benefit from adding multiple
Validators to one parameter we can do so at a later point. When validating user input we
call the parameters validator on the user provided value. We can collect multiple errors
and return them all at once. Being able to return multiple errors in a structured way like
JSON instead of a plain string might help with the presentation in the UI