fix: update rabbitmq lib to fix connection issue
Configure it to set the connection name and consumer prefix to im-manager so
we can tell it apart from the im-inspector. Configure the rabbitmq lib to log
using slog under a group to prevent name clashes. We create a root logger
which we will later also use for the rest of the im-manager. Add some attributes
that appear in every log line like the Go version and the im-manager version.