fix: update rabbitmq lib to fix connection issue DEVOPS-366 (#732)
* fix: update rabbitmq lib to fix connection issue

Configure it to set the connection name and consumer prefix to im-manager so
we can tell it apart from the im-inspector. Configure the rabbitmq lib to log
using slog under a group to prevent name clashes. We create a root logger
which we will later also use for the rest of the im-manager. Add some attributes
that appear in every log line like the Go version and the im-manager version.

* chore: update to Go 1.22.2

* chore: use host/podname instead of im-manager

default to im-manager in case we cannot get the hostname. At least for
the RabbitMQ connection name and consumer tag its better than not having
any other stable identifier.

* chore: log environment and classification

so we know which k8s namespace the pod is in

* chore: update Go to 1.22

* chore: do not log goVersion

we can add later on it when we need it

* chore: update rabbitmq-client to latest release

* chore: logging version via BuildInfo does not work

the way we build. Pass it in instead until the Go issue is resolved

* chore: loki does not index our logs

it already adds labels for the app, pod name and namespace. It will
then show the log line as it is. Lets use a text handler for now
as its easier to read for us humans.