chore: log request id inside Gin context DEVOPS-170 (#735)
* chore: set GIN_MODE to release

We need to make sure every im-manager log line uses the key=val
format so we can parse it in Loki using logfmt. Gin prints some debug
info and a warning in a non logfmt at the start of im-manager. Setting
the mode to release stops this

* chore: return http.request.id to users receiving HTTP 500

the id is generated by https://github.com/samber/slog-gin and put into the
Gin context. Once we have configured Loki to extract that id and index it we
can find all logs to the users failed request.

* chore: log request id when logging inside Gin context

* chore: log using slog in authorization middleware

* chore: address comments