chore: let Gorm log using slog DEVOPS-391 (#741)
* chore: let Gorm log using slog

* chore: log user in Gorm logs when in context

* chore: remove this commit later

this is to test the gorm logs

* chore: put db logs under group

* chore: thread context.Context through user related services

* chore: increase slow query timeout

* chore: remove slog groups

there is no easy/low effort way for us to have a couple of root
level keys like a requestId while putting other slog attributes
in a group. We do not log that much and do not have that many components
that are logging. Lets see how far we get without groups. This way
we can have the 'id', 'user' attribute and extract it using Loki