chore: log user if in context DEVOPS-393 (#743)
* chore: replace samber/slog-gin

we cannot get the user or request ID out of the
c.Request.Context(). This means we cannot add the user
to the log created by samber/slog-gin as our slog.Handler
has no access to it. We therefore write our own Gin
middleware that logs the request and also adds attributes
we put into the gin.Context

* chore: make clear what the id represents