chore: use offset, offset spec next and Last-Event-ID
SSE clients always get the next published message (RabbitMQ offset spec
next) if they do not provide the Last-Event-ID HTTP header. The server
sends the RabbitMQ offset as the SSE Event ID which on disconnect is
sent to the server by the SSE client. This way when the connection is
re-established the SSE client resumes consuming where they left off.