chore: use single docker compose up (#792)
* fix: gitignore for access_token_cache files

these files are named .access_token_cache
so don't end in _

* chore: return an error if we cannot ping redis

* chore: use single compose up

We now fail when redis cannot be reached on startup. So
make it a prod dependency. Improve the RabbitMQ healthcheck
so it is ready to take connections. Rely on Docker healthchecks
and use a single docker compose up to start the smoke-test

* chore: update Go to 1.22.4

there is no alpine-3.18 for this Go version
so we have to update it as well

* chore: review comments

* ci: go back to gha on master