chore: validate CORS config
as it will panic otherwise. A panic in Gins server.GetEngine
was not visible to us when deploying. I mostly saw exit code 2
in k8s which turns out to come from a panic.

Validate the CORS config so we can return an error from main
instead of panicking. Recover from a panic in main so that this
won't happen again.