test: remove commented out mock based tests (#931)
These tests are never going to come back in this form. They constantly pop up
when I grep for code which is just a distraction. They are surely not even
compilable anymore as the code has and will change. We have integration tests
for these packages that cover handlers, services and repositories. We should
strive to make new code unit testable where possible without the need for
mocking. Otherwise, write interation tests