chore: pass context to Gorm DEVOPS-397 (#928)
* chore: use context aware gorm methods

* chore: use correlationId key instead of id

for our correlation id that can come from HTTP or a RabbitMQ message

* docs: fix comment

* chore: remove duplicate log attribute setting

Our slog.ContextHandler adds user and correlationId. The Gin middleware
does not need to do that. Make sure tests fail when we do not set the
user and correlationId as well as forget to register the Gin middleware

* test: fix test race

* chore: fix review comments